+++ RadixWallet/Features/FactorSourceAccess/FactorSourceAccess+View.swift
@@ -1,29 +1,17 @@
+import SwiftUI
+
 // MARK: - FactorSourceAccess.View
-
 extension FactorSourceAccess {
-	struct ViewState: Equatable {
-		let title: String
-		let message: String
-		let externalDevice: String?
-		let isRetryEnabled: Bool
-		let height: CGFloat
-	}
-
-	@MainActor
 	struct View: SwiftUI.View {
-		private let store: StoreOf<FactorSourceAccess>
-
-		init(store: StoreOf<FactorSourceAccess>) {
-			self.store = store
-		}
+		let store: StoreOf<FactorSourceAccess>
 
 		var body: some SwiftUI.View {
-			WithViewStore(store, observe: \.viewState, send: { .view($0) }) { viewStore in
-				content(viewStore)
+			WithPerceptionTracking {
+				content
 					.withNavigationBar {
-						viewStore.send(.closeButtonTapped)
+						store.send(.view(.closeButtonTapped))
 					}
-					.presentationDetents([.fraction(viewStore.height), .large])
+					.presentationDetents([.fraction(store.height), .large])
 					.presentationDragIndicator(.visible)
 					.interactiveDismissDisabled()
 					.presentationBackground(.blur)
@@ -34,25 +22,24 @@ extension FactorSourceAccess {
 			}
 		}
 
-		@ViewBuilder
-		private func content(_ viewStore: ViewStoreOf<FactorSourceAccess>) -> some SwiftUI.View {
+		private var content: some SwiftUI.View {
 			VStack(spacing: .medium3) {
-				Image(asset: AssetResource.signingKey)
+				Image(.signingKey)
 					.foregroundColor(.app.gray3)
 
-				Text(viewStore.title)
+				Text(store.title)
 					.textStyle(.sheetTitle)
 					.foregroundColor(.app.gray1)
 
-				Text(LocalizedStringKey(viewStore.message))
+				Text(LocalizedStringKey(store.message))
 					.textStyle(.body1Regular)
 					.foregroundColor(.app.gray1)
 
-				externalDevice(viewStore.externalDevice)
+				description
 
-				if viewStore.isRetryEnabled {
+				if store.isRetryEnabled {
 					Button {
-						viewStore.send(.retryButtonTapped)
+						store.send(.view(.retryButtonTapped))
 					} label: {
 						Text(L10n.Common.retry)
 							.textStyle(.body1Header)
@@ -67,22 +54,26 @@ extension FactorSourceAccess {
 		}
 
 		@ViewBuilder
-		private func externalDevice(_ value: String?) -> some SwiftUI.View {
-			if let value {
-				HStack(spacing: .medium3) {
-					Image(asset: AssetResource.signingKey)
+		private var description: some SwiftUI.View {
+			if let label = store.label {
+				HStack(spacing: .medium2) {
+					Image(store.kind.icon)
 						.resizable()
 						.frame(.smallest)
 						.foregroundColor(.app.gray3)
 
-					Text(value)
+					Text(label)
 						.textStyle(.secondaryHeader)
 						.foregroundColor(.app.gray1)
 						.padding(.trailing, .small2)
+
+					Spacer()
 				}
 				.padding(.medium2)
 				.background(Color.app.gray5)
-				.cornerRadius(.large1)
+				.cornerRadius(.small1)
+			} else {
+				ProgressView()
 			}
 		}
 	}
@@ -101,10 +92,10 @@ private extension StoreOf<FactorSourceAccess> {
 private extension View {
 	func destinations(with store: StoreOf<FactorSourceAccess>) -> some View {
 		let destinationStore = store.destination
-		return noP2PLinkAlert(with: destinationStore)
+		return errorAlert(with: destinationStore)
 	}
 
-	private func noP2PLinkAlert(with destinationStore: PresentationStoreOf<FactorSourceAccess.Destination>) -> some View {
-		alert(store: destinationStore.scope(state: \.noP2PLink, action: \.noP2PLink))
+	private func errorAlert(with destinationStore: PresentationStoreOf<FactorSourceAccess.Destination>) -> some View {
+		alert(store: destinationStore.scope(state: \.errorAlert, action: \.errorAlert))
 	}
 }
